Adds useProducts hook for fetching products from API
This commit introduces a new file, `use-products.tsx`, which contains a custom hook called `useProducts`. This hook uses the `react-query-kit` library to fetch a list of products from the API endpoint `/v1/products`. The fetched data is then mapped to the `Product` type defined in the file. The hook is created using the `createQuery` function and is configured to fetch the products when called. This addition will enable easy retrieval of products data in other parts of the application.